+++ app/views/products/paycheck.html.haml
@@ -13,7 +13,7 @@
             商品名
           .purchase-content__item__box__show__label
             .purchase-content__item__box__show__label__price
-              ¥1000
+              = @product.price
             .purchase-content__item__box__show__label__tax
               （税込）送料込み
     .purchase-content__info
