+++ app/views/products/show.html.haml
@@ -68,7 +68,7 @@
       = link_to "編集画面にすすむ", edit_product_path, class: "content-detail__form__edit-btn"
       = link_to "商品を削除する", @product, method: :delete, class: "content-detail__form__destroy"
     - else
-      = link_to "購入画面にすすむ",purchase_index_path, class: 'content-detail__form__bay-btn'
+      = link_to "購入画面にすすむ",paycheck_product_path, class: 'content-detail__form__bay-btn'
     .content-detail__form__comment
       = @product.description
     .content-detail__form__icon
