--- app/views/purchase/done.html.haml
@@ -1,3 +1,0 @@
-%h2 購入が完了しました！
-%p テストbook
-%p ¥135,000(送料込み)
